Adding locks
Adds locks as a steel sheet recipe that can be used on any /obj/structure/simple_door to create locked doors.

Uses the machinery var/uid to generate new uid and identify which key belongs to which door.

Locks can be applied to doors that are unlocked with locks to make them be the same lock.

Keys drop after you apply them to a door, meaning there will be a number of keys equal to the number of applied locks.

This is probably not the optimal way to go about things and is a pile of if/else statements, but here we go.

Inexplicably my compile hates soulstones and a lot of wizard/mecha stuff.